PRODEGY_5_Neural_Style_Transfer_ProdigyInfoTech/Application/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
jinja2
python-multipart
pillow-avif-plugin